backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==2.4.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
comm==0.2.0
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
executing==2.0.1
fastapi==0.116.1
fastjsonschema==2.21.2
fonttools==4.61.0
fqdn==1.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.27.1
ipython==8.18.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kiwisolver==1.4.9
lark==1.3.1
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.6
mistune==3.1.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.5.8
notebook==7.5.0
notebook_shim==0.2.4
numpy==2.3.2
openai==1.99.5
packaging==23.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.3
pillow==12.0.0
platformdirs==4.1.0
prometheus_client==0.23.1
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.17.2
pyparsing==3.2.5
python-dateutil==2.8.2
python-dotenv==1.2.1
python-json-logger==4.0.0
pytz==2025.2
pywin32==306
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==25.1.2
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.9.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.47.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.4
tqdm==4.67.1
traitlets==5.14.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.12
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
